requirements.txt
requests
pandas
openpyxl
fastapi
uvicorn
streamlit
sqlalchemy
psycopg2-binary
python-dotenv